(GH-56) Replace usage of Get-ToolPath
This is now making a reasonably huge assumption that NuGet is available
on the System Path, which is what I think Get-ToolPath was at least
partially doing.  I think this is a valid assumption though.
Ultimately, this task needs to be replaced with something that works for
all OSes, so this interim solution should help move things along.